+++ rest/aggs/aggs.go
@@ -1,4 +1,4 @@
-package aggregates
+package aggs
 
 import (
 	"context"
