+++ rest/polygon.go
@@ -1,20 +1,20 @@
 package polygon
 
 import (
-	"github.com/polygon-io/client-golang/rest/aggregates"
+	"github.com/polygon-io/client-golang/rest/aggs"
 	"github.com/polygon-io/client-golang/rest/client"
 )
 
 // todo: add comments for godoc
 
 type polygonClient struct {
-	Aggregates *aggregates.Client
+	Aggregates *aggs.Client
 	// todo: Trades, Quotes, etc
 }
 
 func New(config client.HTTPBaseConfig) *polygonClient {
 	c := client.New(config)
 	return &polygonClient{
-		Aggregates: &aggregates.Client{HTTPBase: c},
+		Aggregates: &aggs.Client{HTTPBase: c},
 	}
 }
